Combine console and network program into one
