+++ console/.gitignore
@@ -2,3 +2,4 @@ ioctl.h
 console
 network
 test
+combined
